+++ scripts/generate-rss.mjs
@@ -6,7 +6,7 @@ import { escape } from "./html-escaper.mjs";
 import { allBlogs } from "../.contentlayer/generated/Blog/_index.mjs";
 import siteMetadata from "../data/siteMetadata.mjs";
 import { getAllSeries, getAllTags } from "../src/lib/get-blog-info.mjs";
-import seriesData from "../data/seriesData.mjs";
+import seriesData from "../data/series-data.mjs";
 
 const generateRssItem = (post) => `
   <item>
